retornando ao styled componet
