+++ server.js
@@ -6,12 +6,6 @@ app.use('/',express.static(resolve(__dirname,"./build")));
 app.use('/acessibilidade',express.static(resolve(__dirname,"./build")));
 app.use('/equipamentos',express.static(resolve(__dirname,"./build")));
 app.use('/tenis',express.static(resolve(__dirname,"./build")));
-app.use('/sig/cadastro-de-cliente',express.static(resolve(__dirname,"./build")));
-app.use('/sig/consulta-de-cliente',express.static(resolve(__dirname,"./build")));
-app.use('/sig/cadastro-de-produto',express.static(resolve(__dirname,"./build")));
-app.use('/sig/consulta-de-produtos',express.static(resolve(__dirname,"./build")));
-app.use('/sig/atualizar-cliente',express.static(resolve(__dirname,"./build")));
-app.use('/sig/atualizar-produto',express.static(resolve(__dirname,"./build")));
 app.use('*',express.static(resolve(__dirname,"./build")));
 app.listen(process.env.PORT || 3000, (err) => {
     if (err) {
