+++ src/admin/dashboard/Administrador/CadastrarAdministrador/styles.css
@@ -1,0 +1,14 @@
+.main{
+    width: 100%;
+    min-height: 600px;
+}
+.exibe-titulo{
+    margin: 2%;
+    text-align: center;
+    font-size: 28px;
+}
+
+#menssagem-erro{
+    color:brown;
+    font-size: 20x;
+}
