+++ src/admin/dashboard/Administrador/ConsultarAdministrador/index.tsx
@@ -1,13 +1,7 @@
 import React from "react";
 import styled from "styled-components";
-import Footer from "../../../../Components/Footer";
-import Cabecalho from "../../../../Components/Menu/Header";
-
-import TabelaCliente from "./tabela";
-const Main = styled.main`
-    width: 100%;
-    min-height: 600px;
-`;
+import TabelaAdimistrador from "./tabela";
+const Main = styled.main``;
 const ExibeTitulo = styled.h3`
     margin: 2%;
     text-align: center;
@@ -49,66 +43,59 @@ const TableExibe = styled.table`
     }
     
     @media screen and (max-width: 1144px) {
-        width: auto;
-        position: static;
-        overflow-x: auto;
-        width: 100%;
-        display: block;
-        padding: 20px .625em .625em .625em;
+       
+    width: auto;
+    position: static;
+    overflow-x: auto;
+    width: 100%;
+    display: block;
+    padding: 20px .625em .625em .625em;
     
-        thead, tbody, thead, th {
-        display: block;
-        }
-        thead {
-        float: left;
-        }
-        tbody {
-        width: auto;
-        position: static;
-        overflow-x: auto;
-        }
-        td, th {
-        padding: 20px .625em .625em .625em;
-        height: 60px;
-        vertical-align: middle;
-        box-sizing: border-box;
-        overflow-x: hidden;
-        overflow-y: auto;
-        width: 120px;
-        font-size: 13px;
-        text-overflow: ellipsis;
-        }
-        thead th {
-        text-align: left;
-        border-bottom: 1px solid #f7f7f9;
-        }
-        tbody tr {
-        display: table-cell;
-        }
-        tbody td {
-        display: block;
-        text-align: center;
-        }
-        tr:nth-child(even) {
-        background: transparent;
-        }
-        tr td:nth-child(odd) {
-        background: #F8F8F8;
-        border-right: 1px solid #E6E4E4;
-        }
-        tr td:nth-child(even) {
-        border-right: 1px solid #E6E4E4;
-        }
+    thead, tbody, thead, th {
+    display: block;
+    }
+    thead {
+    float: left;
+    }
+    tbody {
+    width: auto;
+    position: static;
+    overflow-x: auto;
+    }
+    td, th {
 
+    height: 60px;
+    vertical-align: middle;
+    box-sizing: border-box;
+    overflow-x: hidden;
+    overflow-y: auto;
+    width: 100px;
+    font-size: 12px;
+    }
+    tbody tr {
+    display: table-cell;
+    }
+    tbody td {
+    display: block;
+    text-align: center;
+    }
+    tr:nth-child(even) {
+    background: transparent;
+    }
+
+    tr td:nth-child(even) {
+    border-right: 1px solid #493e3e;
+    }
+    img {
+        width: 100%;
+        height: auto;
+    }
     }
 `;
+const ConsultaAdimistrador = () =>{
 
-
-const ConsultaCliente = () => {
-    return (
-        <>
-            <Cabecalho />
-            <Main>
+    return(
+        <Main id="main">
                 <ExibeTitulo id="exibe-titulo" className="exibe-titulo">Lista de Cliente</ExibeTitulo>
                 <PainelBody id="panel-body" className="panel-name">
 
@@ -128,14 +115,12 @@ const ConsultaCliente = () => {
                             </thead>
 
                             <tbody>
-                             <TabelaCliente/>
+                             <TabelaAdimistrador/>
                             </tbody>
                         </TableExibe>
                     </div>
                 </PainelBody>
             </Main>
-            <Footer />
-        </>
-    );
+    )
 }
-export default ConsultaCliente;
+export default ConsultaAdimistrador
