+++ src/admin/dashboard/Administrador/ConsultarAdministrador/tabela.tsx
@@ -2,52 +2,20 @@ import React, { useEffect, useState } from "react";
 import ICliente from "../../../../utils/interfaces/ICliente";
 import apiFullSports from "../../../../api/apiFullSports";
 import styled from "styled-components";
-import { Box, Button, Modal } from "@mui/material";
-const BtnExibeGroup = styled.div`
-    display: grid;
-    grid-template-columns: repeat(2, auto);
-    grid-auto-rows: minmax(auto, auto);
-    grid-gap: 2px;
 
+const Icone = styled.div`
+    background-color: white;
+    height: 50px;
+    width: 50px;
+    border-radius: 100px;
+    display: flex;
     justify-content: center;
-    margin-left: auto;
+    align-items: center;
     margin-right: auto;
-    padding: 5px;
-`;
-const BtnExibe = styled.button`
-    border: none;
-	border-radius: 5px;
-    background-color: #313131;
-    color: rgb(243, 243, 243);
-    font-size: 14px;
-    width: 70px;
-    height: 35px;
-    :hover{
-        text-decoration: none;
-        background-color: rgb(0, 0, 0);
-        transform:translate(0.3s);
-    }
-    @media screen and (max-width: 1144px) {
-        width: 40px;
-        justify-content: center;
-    }
-`;
-const estiloMenssagem = {
-    position: 'absolute' as 'absolute',
-    top: '50%',
-    left: '50%',
-    transform: 'translate(-50%, -50%)',
-    width: 400,
-    bgcolor: 'background.paper',
-    border: '2px solid #000',
-    boxShadow: 24,
-    pt: 2,
-    px: 4,
-    pb: 3,
-};
-
-
-const TabelaCliente = () => {
+    margin-left: auto;
+    font-size: 22px;
+`
+const TabelaAdimistrador = () => {
     const [open, setOpen] = React.useState(false);
     const handleOpen = () => {
         setOpen(true);
@@ -75,11 +43,15 @@ const TabelaCliente = () => {
        {
         
         clientes.map(item => {
-            if (item.imagemPerfil == null) {
+            if (item.imagemPerfil == null || item.imagemPerfil == undefined) {
                 return (
                     <tr key={item._id.toString()}>
 
-                        <th>sem foto de perfil</th>
+                        <th>
+                       <Icone className="icone">
+                        <p className="text-black">{item.nome.charAt(0)}</p>
+                       </Icone>
+                        </th>
 
                         <th>{`${item.dataCadastro.toLocaleString()}`}</th>
                         <th>{item.cpf}</th>
@@ -88,7 +60,7 @@ const TabelaCliente = () => {
                         <th>{item.sexo}</th>
                         <th>{item.cep}</th>
                         <th>{item.endereco}</th>
-                        <td>
+                        {/* <td>
                             <BtnExibeGroup id="btn-exibe-group" className="btn-exibe-group">
                                 <a href={`/sig/atualizar-cliente/${item._id}`} >
                                     <BtnExibe id="btn-exibe" className="btn-exibe"> Editar </BtnExibe></a>
@@ -102,7 +74,7 @@ const TabelaCliente = () => {
                                         aria-labelledby="child-modal-title"
                                         aria-describedby="child-modal-description"
                                     >
-                                        <Box sx={{ ...estiloMenssagem, width: 650, display: 'flex', justifyContent: 'center' }}>
+                                        <Box id="menssageAlert">
                                             <h2 id="child-modal-title">Deseja mesmo excluir o cliente {item.nome} ?</h2>
                                             <Button onClick={() => deletar(item)} variant="outlined" color="error" >Excluir</Button>
                                             <Button onClick={handleClose} variant="outlined" >Cancelar</Button>
@@ -110,7 +82,7 @@ const TabelaCliente = () => {
                                     </Modal>
                                 </React.Fragment>
                             </BtnExibeGroup>
-                        </td>
+                        </td> */}
                     </tr>
                 )
             } else {
@@ -126,7 +98,7 @@ const TabelaCliente = () => {
                         <th>{item.sexo}</th>
                         <th>{item.cep}</th>
                         <th>{item.endereco}</th>
-                        <td>
+                        {/* <td>
                             <BtnExibeGroup id="btn-exibe-group" className="btn-exibe-group">
                                 <a href={`/sig/atualizar-cliente/${item._id}`} >
                                     <BtnExibe id="btn-exibe" className="btn-exibe"> Editar </BtnExibe></a>
@@ -140,7 +112,7 @@ const TabelaCliente = () => {
                                         aria-labelledby="child-modal-title"
                                         aria-describedby="child-modal-description"
                                     >
-                                        <Box sx={{ ...estiloMenssagem, width: 650, display: 'flex', justifyContent: 'center' }}>
+                                        <Box id="menssageAlert">
                                             <h2 id="child-modal-title">Deseja mesmo excluir o cliente {item.nome} ?</h2>
                                             <Button onClick={() => deletar(item)} variant="outlined" color="error" >Excluir</Button>
                                             <Button onClick={handleClose} variant="outlined" >Cancelar</Button>
@@ -148,7 +120,7 @@ const TabelaCliente = () => {
                                     </Modal>
                                 </React.Fragment>
                             </BtnExibeGroup>
-                        </td>
+                        </td> */}
                     </tr>
                     </>
                 )
@@ -159,4 +131,4 @@ const TabelaCliente = () => {
 
     
 }
-export default TabelaCliente;
+export default TabelaAdimistrador;
