--- src/admin/dashboard/Atualizacoes/Administrador/index.tsx
@@ -1,11 +1,0 @@
-import React from "react"
-import "./styles.css"
-
-
-const AtualizarAdministrador = () => {
-    return (
-        <>
-        
-        </>
-    )};
-export default AtualizarAdministrador
