--- src/admin/dashboard/Cadastros/Administrador/styles.css
@@ -1,3 +1,0 @@
-#menssagem-erro{
-    font-size: 20x;
-}
